Create FunctionChart.md
Each different function gives a different output. The type function identifies the datatype, the int function converts the datatype of the argument to an int, the str() converts the datatype of the argument to an str, the bool() converts the datatype of the argument to an boolean, and the float() converts the datatype of the argument to an float.